Some cleanup on java side
