+++ cpp/src/pypowsybl-api.h
@@ -68,7 +68,6 @@ typedef struct contingency_result_struct {
 
 typedef enum {
     BUS = 0,
-    BRANCH,
     LINE,
     TWO_WINDINGS_TRANSFORMER,
     THREE_WINDINGS_TRANSFORMER,
