+++ java/src/main/java/com/powsybl/dataframe/loadflow/validation/TwtValidationData.java
@@ -1,0 +1,108 @@
+package com.powsybl.dataframe.loadflow.validation;
+
+import com.powsybl.iidm.network.Branch;
+
+/**
+ * @author Yichen TANG <yichen.tang at rte-france.com>
+ */
+class TwtValidationData {
+    String id;
+    double error;
+    double upIncrement;
+    double downIncrement;
+    double rho;
+    double rhoPreviousStep;
+    double rhoNextStep;
+    int tapPosition;
+    int lowTapPosition;
+    int highTapPosition;
+    double targetV;
+    Branch.Side regulatedSide;
+    double v;
+    boolean connected;
+    boolean mainComponent;
+    boolean validated;
+
+    TwtValidationData(String id, double error, double upIncrement, double downIncrement, double rho, double rhoPreviousStep, double rhoNextStep, int tapPosition, int lowTapPosition, int highTapPosition, double targetV, Branch.Side regulatedSide, double v, boolean connected, boolean mainComponent, boolean validated) {
+        this.id = id;
+        this.error = error;
+        this.upIncrement = upIncrement;
+        this.downIncrement = downIncrement;
+        this.rho = rho;
+        this.rhoPreviousStep = rhoPreviousStep;
+        this.rhoNextStep = rhoNextStep;
+        this.tapPosition = tapPosition;
+        this.lowTapPosition = lowTapPosition;
+        this.highTapPosition = highTapPosition;
+        this.targetV = targetV;
+        this.regulatedSide = regulatedSide;
+        this.v = v;
+        this.connected = connected;
+        this.mainComponent = mainComponent;
+        this.validated = validated;
+    }
+
+    String getId() {
+        return id;
+    }
+
+    double getError() {
+        return error;
+    }
+
+    double getUpIncrement() {
+        return upIncrement;
+    }
+
+    double getDownIncrement() {
+        return downIncrement;
+    }
+
+    double getRho() {
+        return rho;
+    }
+
+    double getRhoPreviousStep() {
+        return rhoPreviousStep;
+    }
+
+    double getRhoNextStep() {
+        return rhoNextStep;
+    }
+
+    int getTapPosition() {
+        return tapPosition;
+    }
+
+    int getLowTapPosition() {
+        return lowTapPosition;
+    }
+
+    int getHighTapPosition() {
+        return highTapPosition;
+    }
+
+    double getTargetV() {
+        return targetV;
+    }
+
+    Branch.Side getRegulatedSide() {
+        return regulatedSide;
+    }
+
+    double getV() {
+        return v;
+    }
+
+    boolean isConnected() {
+        return connected;
+    }
+
+    boolean isMainComponent() {
+        return mainComponent;
+    }
+
+    boolean isValidated() {
+        return validated;
+    }
+}
