+++ java/src/main/java/com/powsybl/python/PyPowsyblApiHeader.java
@@ -314,7 +314,6 @@ public final class PyPowsyblApiHeader {
     @CEnum("element_type")
     public enum ElementType {
         BUS,
-        BRANCH,
         LINE,
         TWO_WINDINGS_TRANSFORMER,
         THREE_WINDINGS_TRANSFORMER,
